pyproject.toml
[project]
name = "smart-monitoring"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "matplotlib",
    "pillow",
    "torch",
    "torchvision",
    "jupyter", 
    "tqdm",
    "transformers",
    "evaluate",
    "datasets",
    "accelerate>=0.26.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }